oracle-ai-platform/requirements.txt
# requirements.txt
# Core dependencies


# Oracle database connection
cx_Oracle==8.3.0

# Web framework (choose one)
fastapi==0.104.1
uvicorn[standard]==0.24.0
# OR for Streamlit
streamlit==1.28.0

# LLM integration
openai==1.3.0
anthropic==0.7.4
ollama==0.1.2

# Vector database
chromadb==0.4.18
sentence-transformers==2.2.2

# Data processing
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2

# PDF/text processing
PyPDF2==3.0.1
python-docx==1.1.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Optional: for PDF extraction if using alternative
pdfplumber==0.10.3